Move helper functions from _inject.ts to _util.ts
