+++ js/_copy-as-json.ts
@@ -1,4 +1,4 @@
-import { $ } from './_inject';
+import { $ } from './_util';
 
 $('#copy')?.addEventListener('click', () =>
     navigator.clipboard.writeText($('#data').innerText));
