+++ js/hole.tsx
@@ -1,4 +1,4 @@
-import { $, $$, comma } from './_inject';
+import { $, $$, comma } from './_util';
 import CodeMirror                from './_codemirror-legacy';
 import                                './_copy-as-json';
 import diffTable                 from './_diff';
